Use unsigned int for port numbers.

+++ inc/WireCellIface/IDataFlowGraph.h
@@ -18,7 +18,7 @@ namespace WireCell {
 	/// port number to given head port number Return false on
 	/// error.
 	virtual bool connect(INode::pointer tail, INode::pointer head,
-			     int tail_port=0, int head_port=0) = 0;
+			     size_t tail_port=0, size_t head_port=0) = 0;
 
 	/// Run the graph, return false on error.
 	virtual bool run() = 0;
